refactor: split monolithic cli.py into cli/ package
Closes #71

- Replace 2504-line cli.py with cli/ package (23 modules)
- cli/__init__.py re-exports main() and _apply_config_defaults
- cli/_main.py contains main() with argument parser setup
- cli/_helpers.py contains shared helper functions
- cli/_config.py contains config-related helpers
- Each subcommand in its own module (_analyze, _export, etc.)
- All 691 tests pass unchanged
- No behavioral changes — pure refactor
